Removed some comments from lcats, changed GIT_STRATEGY to none for all jobs except tpl, created a spheralutils.py in scripts so all python scripts use the same sexe routine
